Fixed bugs on finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
-monthly_income = int(input("Enter your monthly income: "))
-monthly_expense = int(input("Enter your monthly expensed: "))
-monthly_savings = monthly_income -monthly_expense
+monthly_income = input("Enter your monthly income: ")
+monthly_expense = input("Enter your monthly expensed: ")
+monthly_savings = float(monthly_income) - float(monthly_expense)
 
 projected_savings = monthly_saving * 12 + (monthly_saving * 12 * 0.05)
 print(f'Your monthly savings are ${monthly_saving}')
